+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/product/web/bodies/ProductResponse.java
@@ -11,7 +11,7 @@ public class ProductResponse {
     private final String description;
     private final int amount;
     private final String unit;
-    private final int price;
+    private final double price;
 
     public ProductResponse(Product product) {
         this.id=product.getId();
